Moved all screen components to separate files.
